+++ src/modules/functions.test.js
@@ -1,0 +1,1 @@
+const functions = require('./functions.js');
